feat: run upgrade using init API
