+++ internal/app/osd/proto/api.proto
@@ -19,7 +19,6 @@ service OSD {
   rpc Routes(google.protobuf.Empty) returns (RoutesReply) {}
   rpc Stats(StatsRequest) returns (StatsReply) {}
   rpc Top(google.protobuf.Empty) returns (TopReply) {}
-  rpc Upgrade(UpgradeRequest) returns (UpgradeReply) {}
   rpc Version(google.protobuf.Empty) returns (Data) {}
 }
 
@@ -100,9 +99,3 @@ message DFStat {
   uint64 available = 3;
   string mounted_on = 4;
 }
-
-message UpgradeRequest {
-	string url = 1;
-}
-
-message UpgradeReply { string ack = 1; }
